[implement]: Post CRUD 구현

+++ src/main/java/com/yemin/twitter/repository/PostImageRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PostImageRepository extends JpaRepository<PostImageVO, Integer> {
+public interface PostImageRepository extends JpaRepository<PostImageVO, Long> {
 
 }
